backend/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
python-dotenv==0.20.0
tweepy==4.8.0